refactor(web): extract useConversations hook, share test stubs, tidy feedback camelcase
