+++ web/test/shell.test.tsx
@@ -24,52 +24,10 @@ import { ConversationList } from '@/components/shell/conversation-list';
 import { Sidebar } from '@/components/shell/sidebar';
 import { type ConversationRow, db } from '@/lib/db';
 import { useUiStore } from '@/lib/ui-store';
-
-// jsdom lacks ResizeObserver, which the vendored Conversation (use-stick-to-bottom)
-// constructs with `new`. Provide a constructable no-op stub.
-class ResizeObserverStub {
-  callbacks: ResizeObserverCallback[] = [];
-
-  constructor(callback: ResizeObserverCallback) {
-    this.callbacks.push(callback);
-  }
-
-  disconnect(): void {
-    this.callbacks = [];
-  }
-
-  observe(): void {
-    this.callbacks.at(0);
-  }
-
-  unobserve(): void {
-    this.callbacks.at(0);
-  }
-}
-
-// jsdom 26 only exposes localStorage for a concrete origin; under the default
-// vitest jsdom URL the origin is opaque, so we install a conforming in-memory
-// store the transport's getAnonUserId can use.
-const createMemoryStorage = (): Storage => {
-  const store = new Map<string, string>();
-
-  return {
-    clear: () => {
-      store.clear();
-    },
-    getItem: (key) => store.get(key) ?? null,
-    key: (index) => store.keys().toArray()[index] ?? null,
-    get length() {
-      return store.size;
-    },
-    removeItem: (key) => {
-      store.delete(key);
-    },
-    setItem: (key, value) => {
-      store.set(key, value);
-    },
-  };
-};
+import {
+  createMemoryStorage,
+  ResizeObserverStub,
+} from '@/test/helpers/dom-stubs';
 
 beforeAll(() => {
   vi.stubGlobal('ResizeObserver', ResizeObserverStub);
