+++ web/test/user.test.ts
@@ -1,33 +1,10 @@
 import { beforeEach, expect, test, vi } from 'vitest';
 
 import { ANON_USER_ID_KEY, getAnonUserId } from '@/lib/user';
+import { createMemoryStorage } from '@/test/helpers/dom-stubs';
 
 const UUID_V4 =
   /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/u;
-
-// jsdom 26 only exposes localStorage for a concrete origin; under the default
-// vitest jsdom URL the origin is opaque, so we install a conforming in-memory
-// store for the module under test.
-const createMemoryStorage = (): Storage => {
-  const store = new Map<string, string>();
-
-  return {
-    clear: () => {
-      store.clear();
-    },
-    getItem: (key) => store.get(key) ?? null,
-    key: (index) => store.keys().toArray()[index] ?? null,
-    get length() {
-      return store.size;
-    },
-    removeItem: (key) => {
-      store.delete(key);
-    },
-    setItem: (key, value) => {
-      store.set(key, value);
-    },
-  };
-};
 
 beforeEach(() => {
   vi.stubGlobal('localStorage', createMemoryStorage());
